app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["arkady <atyursrvice@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
pydantic = "^2.3.0"
uvicorn = "^0.23.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.test.dependencies]
httpx = "^0.25.0"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
